backend/requirements.txt
# Core FastAPI and server
fastapi==0.111.0
uvicorn==0.29.0
starlette==0.37.2
pydantic==2.11.5
pydantic-settings==2.9.1

# Database
SQLAlchemy==2.0.41
asyncpg==0.30.0
alembic==1.16.1
greenlet==3.2.2
psycopg2-binary==2.9.9

# Celery and Task Queue
celery==5.4.0
redis==5.0.8
billiard==4.2.1
kombu==5.5.4
vine==5.1.0
amqp==5.3.1

# Authentication and Security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
email_validator==2.2.0

# MinIO Object Storage
minio==7.2.8

# --- NEW: Google AI and Image Handling ---
google-generativeai==0.7.1
requests==2.32.3
Pillow==10.4.0

# Utilities
python-dotenv==1.1.0
python-multipart==0.0.20
anyio==4.9.0
httpx==0.28.1
numpy==2.1.3